# Conflicts:
#	backend/src/main/java/com/spring/mmm/domain/mukgroups/domain/MukboEntity.java
#	backend/src/main/java/com/spring/mmm/domain/mukgroups/service/MukgroupServiceImpl.java

+++ backend/src/main/java/com/spring/mmm/domain/mbtis/controller/response/AnswerResponse.java
@@ -7,7 +7,7 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 public class AnswerResponse {
-    private Long answerId;
+    private Integer answerId;
     private String answerContext;
     private String answerImage;
 }
